Auto import of structs and enums
With Thrift 0.9.4 (or 1.0, it depends) we get additional metadata in the
erlang implementation that allows us to auto import structs and
enums. This eliminates a bunch of boilerplate code that we currently
have to write to re-define enums, and tell our struct module which
structs we want to import.

Riffed is much more automatic now, all that needs to be done is the
enumerize_* functions.