Get mock from Hex instead of Github
Previously we were getting the 'mock' package directly from Github at
jjh42/mock. It has a bug that causes a warning to be displayed when retrieving
it that way. It expects a @version to be available in its mix.exs. That
variable is populated if you retrieve it through Hex. So that's what we do now.