Added bang functions to client
The Elixir idiom is to have a function that ends with an exclamation
point throw an exception, and an unadorned function return `{:ok,
response}` or `{:error, reason}`. This PR changes Riffed to respoect
that idiom. It is a breaking change, and will break all clients
upgrading from any version of elixometer < 1.5